requirements.txt
streamlit
langchain
langchain-community
langchain-core
sentence-transformers
faiss-cpu
python-dotenv
numpy
tqdm
tensorflow
tf-keras
Pillow
flask
requests
langchain-google-genai